sportguard-ai/backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
google-cloud-vision
google-cloud-firestore
google-cloud-storage
python-multipart
google-generativeai
python-dotenv
